+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-24 03:11:38
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: ded3d41225a2b486b5435b067da2bd4e241bb591
+
+## [ForwardWidgets-2kuai] - 2025-08-24 03:11:38
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 893b1c876ead48b75e87840ea1b90bd56b9857a3
+
+## [ForwardWidgets-pack1r] - 2025-08-24 03:11:38
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 156c4c5cd7088695ad45b89f3248ebbd06098c37
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-23 02:56:14
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main